Minor font and color adjustments
Use RoyalBlue box color. Looks more accurate with the PLA I have.

Include - CutoutMargin/2 in the rectangular hold edge measurement since
that goes into the font location. Now it is better centered over the
holes.

For all text, fix the m/2 calculations so that adjusting m results in
the text being in the same place. Add or subtract 1 to keep it in nearly
the same place as before (off by 0.1)