Separate PartMargin and CutoutMargin for panel
Make separate PanelPartMargin and PanelCutoutMargin so I can model the
panels with different margins. Intended for use with laser cutters, I
can model the panels with tighter tolerances than the 3d printer. Leave
these as copies of the originals, but add comments to set the margin to
-0.1 for a laser cutter. The laser cutter removes about 0.1 mm from the
edge of the cut, so you have to cut larger than the size you want.